perf(best-media): move getBestMedia into the sequences worker
Defense-in-depth driven by the same fresh-boot evidence as the
species-distribution move: SQLite queries that are fast on warm FS
cache can freeze the main thread on cold disk. The favorites CTE in
getBestMedia scans observations joined with media (2.7M-row join on
gmu8_leuven) and we have no cold-start measurement of it; the
auto-scored CTE has four scan levels + two ROW_NUMBER window functions
and can legitimately take multiple seconds on ML-run datasets with
populated bboxes and millions of observations. Even with the phase-2
short-circuit covering no-bbox datasets, bbox-having datasets remain
at risk of blocking the renderer's UI if the query runs on main.

Move the media:get-best IPC handler off main, same shape as
sequences:get-species-distribution.

Changes:

- Factor the runInWorker helper out of ipc/sequences.js into a new
  src/main/services/sequences/runInWorker.js so media.js can share it
  without duplicating the 20-line Worker bootstrap. Behavior is
  identical: same worker file (out/main/sequences-worker.js, from the
  existing sequences-worker rollup input), same __dirname resolution
  (all of src/main/** flattens into out/main/ at build time).

- Add a 'best-media' task type to src/main/services/sequences/worker.js.
  The worker file name is already broader than "sequences" in practice
  (it handles species-distribution, timeseries, heatmap, daily-activity)
  so adding best-media fits the existing pattern; renaming is
  out-of-scope for this change.

- Update the media:get-best handler in src/main/ipc/media.js to
  dispatch through runInWorker. Drop the now-unused getBestMedia
  import from the handler file.

Trade-off: adds ~1s of Worker-spawn overhead per getBestMedia call.
The BestMediaCarousel query uses staleTime: Infinity (commit 146ae0c),
so this cost is paid once per study per session. In exchange, no UI
freeze is possible from this path regardless of dataset shape,
cold/warm state, or future perf regressions in the auto-scored CTE.

Tests: existing bestMedia.test.js exercises getBestMedia as a pure
function against a test DB, so the Worker migration does not affect
it. Full suite 920/920.
