+++ src/main/ipc/media.js
@@ -11,12 +11,12 @@ import {
   getMediaBboxesBatch,
   checkMediaHaveBboxes,
   getVideoFrameDetections,
-  getBestMedia,
   updateMediaTimestamp,
   updateMediaFavorite,
   countMediaWithNullTimestamps,
   closeStudyDatabase
 } from '../database/index.js'
+import { runInWorker } from '../services/sequences/runInWorker.js'
 
 /**
  * Register all media-related IPC handlers
@@ -91,7 +91,10 @@ export function registerMediaIPCHandlers() {
     }
   })
 
-  // Get best media files scored by bbox quality heuristic
+  // Get best media files scored by bbox quality heuristic.
+  // Runs in the sequences worker so the favorites CTE and (potentially heavy)
+  // auto-scored CTE don't block the renderer's UI on cold-cache runs. The
+  // bbox short-circuit inside getBestMedia keeps the no-bbox case cheap too.
   ipcMain.handle('media:get-best', async (_, studyId, options = {}) => {
     try {
       const dbPath = getStudyDatabasePath(app.getPath('userData'), studyId)
@@ -100,7 +103,12 @@ export function registerMediaIPCHandlers() {
         return { error: 'Database not found for this study' }
       }
 
-      const bestMedia = await getBestMedia(dbPath, options)
+      const bestMedia = await runInWorker({
+        type: 'best-media',
+        dbPath,
+        studyId,
+        options
+      })
       return { data: bestMedia }
     } catch (error) {
       log.error('Error getting best media:', error)
